requirements.txt
flask==2.3.3
google-generativeai==0.3.2
chromadb==0.4.18
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.1
tqdm==4.66.1
playwright==1.40.0
